whitelistTransaction API enhancement, AccountNotFoundError with build transaction
* whitelistTransaction now accept both networkId and network_id as param,
in order to align to python sdk while keep backward compatibility,
using string is not allowed anymore as it wasn't clear anyway, and docs instruct to
use param object.
the type "envelop" also removed as it wasn't used, android already fixed it
and it was specific to tests whitelist service that anyway uses python code.
* Build account when account doesn't exist will yield an AccountNotFoundError
instead of js-kin-sdk error.
* tests added to all of those scenarios
* KinAccount unit tests was restructure to separate topics for better readability,
fixed retry in post requests (w/wo redirect), added and reorder tests with retry for better coverage.